requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
google-generativeai
pymupdf
python-docx
python-dotenv
aiofiles
